Change IgnoreDuplicateNavThreshold type from int to long.
This threshold is converted to base::TimeDelta in the Chromium, which
uses a int64_t to represent microseconds. Also, following Android API
guidelines for durations, `long` is the recommended type for such
values. So we should use long here instead of int to remove unnecessary
conversion.

Bug: 489977120
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7780889
Commit-Queue: Anna Sato <annasato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1618890}

+++ android_webview/java/src/org/chromium/android_webview/AwSettings.java
@@ -204,7 +204,7 @@ public class AwSettings {
     private int mMixedContentMode = WebSettings.MIXED_CONTENT_NEVER_ALLOW;
     private int mAttributionBehavior = AttributionBehavior.APP_SOURCE_AND_WEB_TRIGGER;
     private boolean mIgnoreDuplicateNavEnabled;
-    private int mIgnoreDuplicateNavThresholdMs = -1;
+    private long mIgnoreDuplicateNavThresholdMs = -1;
 
     @SpeculativeLoadingAllowedFlags
     private int mSpeculativeLoadingAllowedFlags =
@@ -835,7 +835,7 @@ public class AwSettings {
      *
      * @param thresholdMs the threshold in milliseconds.
      */
-    public void setIgnoreDuplicateNavThreshold(int thresholdMs) {
+    public void setIgnoreDuplicateNavThreshold(long thresholdMs) {
         synchronized (mAwSettingsLock) {
             if (mIgnoreDuplicateNavThresholdMs != thresholdMs) {
                 mIgnoreDuplicateNavThresholdMs = thresholdMs;
@@ -845,7 +845,7 @@ public class AwSettings {
     }
 
     @CalledByNative
-    public int getIgnoreDuplicateNavThreshold() {
+    public long getIgnoreDuplicateNavThreshold() {
         synchronized (mAwSettingsLock) {
             return mIgnoreDuplicateNavThresholdMs;
         }
@@ -2113,7 +2113,7 @@ public class AwSettings {
     }
 
     @CalledByNative
-    private int getIgnoreDuplicateNavThresholdLocked() {
+    private long getIgnoreDuplicateNavThresholdLocked() {
         assert Thread.holdsLock(mAwSettingsLock);
         return mIgnoreDuplicateNavThresholdMs;
     }
